Task2/apz-pzpi-21-4-sorokin-vadym-task2/ai-server/pyproject.toml
[tool.poetry]
name = "ai-server"
version = "0.1.0"
description = ""
authors = ["nicourrrn <s57111702@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
psycopg = {extras = ["binary"], version = "^3.1.19"}
pyjwt = "^2.8.0"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
